[quantizer] merged consumer.go with sql.go
